# wire: optimize generation of Version Negotiation packets