```
name                                old time/op    new time/op    delta
ComposeVersionNegotiationPacket-16     900ns ± 2%     317ns ± 3%  -64.74%  (p=0.000 n=10+10)

name                                old alloc/op   new alloc/op   delta
ComposeVersionNegotiationPacket-16      120B ± 0%       64B ± 0%  -46.67%  (p=0.000 n=10+10)

name                                old allocs/op  new allocs/op  delta
ComposeVersionNegotiationPacket-16      6.00 ± 0%      2.00 ± 0%  -66.67%  (p=0.000 n=10+10)